requirements.txt
flask
python-escpos[win32]
pywin32
python-dotenv
Pillow
streamlit
requests
